requirements.txt
functions-framework==2.1.2
google-api_core==1.26.3
google-cloud-bigquery-reservation==1.1.0
